The call to mDetector.release() in FrameProcessingRunnable.release() has been removed. This is due to the fact that, as it's called in releaseCamera(), which invokes in CamView.stop() which happens whenever the user leaves the app at ALL
The call to Detector.release() removes the processor as well. The only other way to mitigate this
would be passing the processor all the way down the chain to the FPR and let it attach it in the case that a user leaves the app, comes back, and still wants to detect stuff. Otherwise, the
detector will break anytime the user leaves the app!

Best suggestion: Pass the processor down the chain to the FPR, this allows us to release the resources fully EVERYTIME the user leaves the app. Then when the FPR is reactivated it can reattach the processor.

I'll look into doing it this way, if this pull is accepted. That'l be a change I add during finalization. But for now, for the purposes of my app, I'll make this simple change just for the sake of time.

Signed-off-by: IOnlyFetchBranches <demarcus-joachim@georgiasouthern.edu>