removes updates to display orientation in SurfaceViewContainer
The preview size is already transposed to account for display/device orientation when it
is passed to the surface view container so there is no need to further modify it.

There was also a condition in landscape configuration where the preview was never
set, which would cause the layout code to greatly distort the display by, in some cases,
basing the layout solely on the width/height of the surface view container.